Update partition_data.m
If e1 is called as e2 (e.g., findfield(data,evt1), e1 was a giving a struct array with fields. Don't really understand why because evt1 was = data_start and there was a field data_start, as for evt2 (data_end). I changed the formulation (not very clean) but it work.